AgentHost - 💄 extract the IAgentHostGitService into a separate file
