+++ src/vs/platform/agentHost/test/common/sessionTestHelpers.ts
@@ -171,7 +171,7 @@ export function encodeString(text: string): Uint8Array {
  * exercise the {@link AgentService} but don't care about git state.
  * Tests that DO care about git state should pass their own implementation.
  */
-export function createNoopGitService(): import('../../node/agentHostGitService.js').IAgentHostGitService {
+export function createNoopGitService(): import('../../common/agentHostGitService.js').IAgentHostGitService {
 	return {
 		_serviceBrand: undefined,
 		isInsideWorkTree: async () => false,
