+++ src/vs/platform/agentHost/test/node/agentHostChangesetService.test.ts
@@ -17,7 +17,7 @@ import { ChangesetStatus, SessionStatus, withSessionGitState, type Changeset } f
 import { AgentHostChangesetService } from '../../node/agentHostChangesetService.js';
 import { IAgentHostChangesetSubscriptionService } from '../../common/agentHostChangesetSubscriptionService.js';
 import { NULL_CHECKPOINT_SERVICE } from '../../common/agentHostCheckpointService.js';
-import { IAgentHostGitService } from '../../node/agentHostGitService.js';
+import { IAgentHostGitService } from '../../common/agentHostGitService.js';
 import { AgentHostStateManager } from '../../node/agentHostStateManager.js';
 import { AgentConfigurationService } from '../../node/agentConfigurationService.js';
 import { SessionDatabase } from '../../node/sessionDatabase.js';
